+++ .idea/dictionaries/vikra.xml
@@ -2,6 +2,7 @@
   <dictionary name="vikra">
     <words>
       <w>chrono</w>
+      <w>chronofit</w>
       <w>gson</w>
       <w>jsoncircuit</w>
       <w>lapbutton</w>
