+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.chrono">
+    package="ca.chronofit.chrono">
 
     <application
         android:icon="@drawable/ic_chrono_app_icon"
