+++ app/src/main/java/ca/chronofit/chrono/util/components/ModalBottomSheet.kt
@@ -1,0 +1,23 @@
+package ca.chronofit.chrono.util.components
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import ca.chronofit.chrono.R
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+
+class ModalBottomSheet : BottomSheetDialogFragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_dashboard_bottom_sheet, container, false)
+    }
+}
