+++ app/src/main/java/ca/chronofit/chrono/util/objects/CircuitObject.kt
@@ -1,0 +1,13 @@
+package ca.chronofit.chrono.util.objects
+
+class CircuitObject {
+    var name: String? = null
+    var sets: Int? = null
+    var work: Int? = null
+    var rest: Int? = null
+    var iconId: Int? = null
+
+    override fun toString(): String {
+        return "Circuit [name: ${this.name}, sets: ${this.sets}, work: ${this.work}, rest: ${this.rest}]"
+    }
+}
