+++ app/src/main/java/ca/chronofit/chrono/util/objects/CircuitsObject.kt
@@ -1,0 +1,8 @@
+package ca.chronofit.chrono.util.objects
+
+import java.io.Serializable
+
+class CircuitsObject : Serializable {
+    var key: String = "CIRCUITS"
+    var circuits: ArrayList<CircuitObject>? = ArrayList()
+}
