+++ app/src/main/java/ca/chronofit/chrono/util/objects/LapObject.kt
@@ -1,0 +1,7 @@
+package ca.chronofit.chrono.util.objects
+
+class LapObject {
+    var lapNum: Int = 0
+    var lapTime: Long = 0
+    var totalTime: Long = 0
+}
