+++ app/src/main/java/ca/chronofit/chrono/util/services/FirebaseCloudMessaging.kt
@@ -1,0 +1,59 @@
+package ca.chronofit.chrono.util.services
+
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import android.util.Log
+import androidx.core.app.NotificationCompat
+import ca.chronofit.chrono.R
+import ca.chronofit.chrono.main.MainActivity
+import com.google.firebase.messaging.FirebaseMessagingService
+import com.google.firebase.messaging.RemoteMessage
+
+class FirebaseCloudMessaging : FirebaseMessagingService() {
+
+    override fun onMessageReceived(remoteMessage: RemoteMessage) {
+        // Required to display notification if the app is open
+        sendNotification(remoteMessage.notification!!)
+    }
+
+    private fun sendNotification(notification: RemoteMessage.Notification) {
+        val intent = Intent(this, MainActivity::class.java)
+        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+        val pendingIntent = PendingIntent.getActivity(
+            this, 0 /* Request code */, intent,
+            PendingIntent.FLAG_ONE_SHOT
+        )
+
+        val channelId = getString(R.string.default_fcm_notification_channel_id)
+        val notificationBuilder = NotificationCompat.Builder(this, channelId)
+            .setSmallIcon(R.drawable.ic_notification_logo)
+            .setContentTitle(notification.title)
+            .setContentText(notification.body)
+            .setAutoCancel(true)
+            .setContentIntent(pendingIntent)
+
+        val notificationManager =
+            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val channel = NotificationChannel(
+                channelId,
+                "fcm notification channel",
+                NotificationManager.IMPORTANCE_DEFAULT
+            )
+            notificationManager.createNotificationChannel(channel)
+        }
+
+        notificationManager.notify(0, notificationBuilder.build())
+    }
+
+    override fun onNewToken(newToken: String) {
+        Log.d("FCM Messaging Service", "Refreshed token: $newToken")
+        super.onNewToken(newToken)
+    }
+
+}
