--- app/src/main/java/com/example/chrono/util/objects/CircuitObject.kt
@@ -1,13 +1,0 @@
-package com.example.chrono.util.objects
-
-class CircuitObject {
-    var name: String? = null
-    var sets: Int? = null
-    var work: Int? = null
-    var rest: Int? = null
-    var iconId: Int? = null
-
-    override fun toString(): String {
-        return "Circuit [name: ${this.name}, sets: ${this.sets}, work: ${this.work}, rest: ${this.rest}]"
-    }
-}
