--- app/src/main/java/com/example/chrono/util/objects/CircuitsObject.kt
@@ -1,8 +1,0 @@
-package com.example.chrono.util.objects
-
-import java.io.Serializable
-
-class CircuitsObject : Serializable {
-    var key: String = "CIRCUITS"
-    var circuits: ArrayList<CircuitObject>? = ArrayList()
-}
