--- app/src/main/java/com/example/chrono/util/objects/LapObject.kt
@@ -1,7 +1,0 @@
-package com.example.chrono.util.objects
-
-class LapObject {
-    var lapNum: Int = 0
-    var lapTime: Long = 0
-    var totalTime: Long = 0
-}
